Add support for function offset in batch serialisation (#187)
With these changes, we can pass as row offset a function, mapping from
the row offset passed to the `.call(n)` method of the batch-serialised
graph to a new row offset. It can be a different function for each load
or store handle.
For example, in the newly added unit test we pass the following remote
handle:
`RemoteHandle(buffer, lambda i: popxl.constant(1))`
Which means that we'll use the row offset 1 regardless of the `n` passed
to `.call()`.