Fix bug in checkpoint loading (#193)
Currently when trying to match variable names to tensors in a checkpoint
a line replaces the "." with "_" in the name, because of some old code
in the checkpoint saving (which has been removed a few months ago). This
means that currently checkpoint loading silently fails.
This PR fixes this issue.