Fp8 best scale op (#190)
This PR introduces to methods to compute the scaling bias when casting
from FP32/FP16 to FP8 (both 1_4_3 and 1_5_2):
- `amax`: finds the absolute max of the input data and uses it to
compute the scaling bias
- `mse`: builds the histogram of the input data and, for each of the
scaling bias integer options, computes the mean-squared error from
underflow, overflow and within range. The scaling bias with the lowest
mse is returned.

The `mse` method only works in hw with FP8 support (e.g. Mk2x) with the
option to enable saturation for overflow instead of returning NaN.

---------

Co-authored-by: Sergio Perez <sergioperezperez24@gmail.com>