Minor changes for compatability with SWQA CI (#208)
Hi - small changes needed to get things working for SWQA CI.

- Changing dataset directory to location on CI with MNIST already
downloaded, lowers chances of intermittent network errors when
downloading. Happy for this to be changed if needed - alternative
implementation would be to check for env var (DATASETS_DIR) but this
feels easiest?

- Marking MNIST tests as serial - both invoke the download and this
caused some issues in testing when they'd work in parallel, hitting some
kind of EOF error.